requirements.txt
Django==3.0.5
Unipath==1.1
asgiref==3.2.7
django-model-utils==4.0.0
djangorestframework==3.12.4
psycopg2>=2.8.6,<2.9
pytz==2022.1
setuptools==49.2.1
sqlparse==0.3.1
